[FIX] fieldservice_account: tests and travis
