+++ fieldservice_account/__init__.py
@@ -1,4 +1,3 @@
 # Copyright (C) 2018 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from . import models
-from .init_hook import pre_init_hook
