+++ fieldservice_account/models/account_invoice_line.py
@@ -7,9 +7,4 @@ from odoo import fields, models
 class AccountInvoiceLine(models.Model):
     _inherit = 'account.invoice.line'
 
-    fsm_order_ids = fields.One2many(
-        comodel_name='fsm.order',
-        inverse_name='invoice_line_id',
-        string='FSM Orders',
-        readonly=True, copy=False,
-    )
+    fsm_order_id = fields.Many2one('fsm.order', string='FSM Order')
