+++ fieldservice_account/readme/CONTRIBUTORS.rst
@@ -1,3 +1,4 @@
 * Michael Allen <mallen@opensourceintegrators.com>
 * Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
 * Raphaël Reverdy <raphael.reverdy@akretion.com>
+* Brian McMaster <brian@mcmpest.com>
