+++ fieldservice_account/views/account_invoice.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
     <record id="account_invoice_form" model="ir.ui.view">
