+++ fieldservice_account/views/fsm_order.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="view_fsm_order_form_inherit_contract" model="ir.ui.view">
+    <record id="view_fsm_order_form_inherit_account" model="ir.ui.view">
         <field name="model">fsm.order</field>
         <field name="inherit_id" ref="fieldservice.fsm_order_form"/>
         <field name="arch" type="xml">
