+++ fieldservice_account_analytic/__init__.py
@@ -3,3 +3,4 @@
 
 from . import models
 from . import wizard
+from .init_hook import pre_init_hook
