+++ fieldservice_account_analytic/tests/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2019 Brian McMaster <brian@mcmpest.com>
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
+
+from . import test_fsm_account_wizard
