--- fieldservice_account_analytic/views/fsm_person.xml
@@ -1,26 +1,0 @@
-<odoo>
-
-    <record id="fsm_person_form" model="ir.ui.view">
-        <field name="name">fsm.person.form</field>
-        <field name="model">fsm.person</field>
-        <field name="inherit_id" ref="fieldservice.fsm_person_form"/>
-        <field name="arch" type="xml">
-            <field name="owner_id" position="after">
-                <field name="customer_id"/>
-            </field>
-        </field>
-    </record>
-
-    <!-- Field Service Location Person Form View-->
-    <record id="fsm_location_person_form_view_analytic" model="ir.ui.view">
-        <field name="name">fsm.location.person.form</field>
-        <field name="model">fsm.location.person</field>
-        <field name="inherit_id" ref="fieldservice.fsm_location_person_form_view"/>
-        <field name="arch" type="xml">
-            <field name="owner_id" position="after">
-                <field name="customer_id"/>
-            </field>
-        </field>
-    </record>
-
-</odoo>
