+++ fieldservice_isp_account/models/__init__.py
@@ -2,8 +2,6 @@
 
 from . import (
     account_invoice_line,
-    analytic_account,
-    fsm_location,
-    fsm_location_person,
     fsm_order,
+    fsm_person,
 )
