+++ fieldservice_isp_account/views/fsm_order.xml
@@ -3,7 +3,8 @@
     <record id="fsm_order_form_isp_account" model="ir.ui.view">
         <field name="name">fsm.order.form.isp.account</field>
         <field name="model">fsm.order</field>
-        <field name="inherit_id" ref="fieldservice.fsm_order_form_account"/>
+        <field name="inherit_id" 
+               ref="fieldservice_account_analytic.fsm_order_form_analytic"/>
         <field name="arch" type="xml">
             <page name="accounting" position="inside">
                 <header>
